CR Changes
Former-commit-id: b9d236f96218a448bba61afd455cd8b7c2d58318

+++ id/src/main/java/com/simprints/id/domain/moduleapi/face/responses/entities/FaceMatchResult.kt
@@ -6,7 +6,7 @@ import kotlinx.android.parcel.Parcelize
 
 @Parcelize
 data class FaceMatchResult(val guidFound: String,
-                           val confidence: Int) : Parcelable
+                           val confidence: Float) : Parcelable
 
 fun IFaceMatchResult.fromModuleApiToDomain() =
     FaceMatchResult(guid, confidence)
